Création Notifications + Register et Order + MAJ migrations et routes

+++ database/migrations/2024_12_08_093620_create_sessions_table.php
@@ -11,14 +11,16 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('sessions', function (Blueprint $table) {
-            $table->string('id', 191)->primary();
-            $table->foreignId('user_id')->nullable()->constrained()->onDelete('cascade');
-            $table->string('ip_address', 45)->nullable();
-            $table->text('user_agent')->nullable();
-            $table->text('payload');
-            $table->integer('last_activity');
-        });
+        if (!Schema::hasTable('sessions')) {
+            Schema::create('sessions', function (Blueprint $table) {
+                $table->string('id', 191)->primary();
+                $table->foreignId('user_id')->nullable()->constrained()->onDelete('cascade');
+                $table->string('ip_address', 45)->nullable();
+                $table->text('user_agent')->nullable();
+                $table->text('payload');
+                $table->integer('last_activity');
+            });
+        }
     }
 
     /**
